Add files via upload
In the virtual mouse stuff I had a lot of problems as, at the time the new input system was being changed, this lead to things breaking and if I'm honest a lot of needless headaches. An off the shelf solution would have worked just fine and would have lead to less wondering how they broke it this time.